Updated run instructions, and cleaned up
README.md has been updated to include new run instructions, i.e. the
script now has to be run with the node server not running.

Also removed some functions out of index.js, as I've actually moved them
to app.js and so that the resetdb function is no longer reachable with a
get request